bot/requirements.txt
# Discord
discord.py==2.3.2

# HTTP client
aiohttp==3.9.3

# Chess board rendering
python-chess==1.999
cairosvg==2.7.1

# Environment
python-dotenv==1.0.1
